% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annulus.r
\name{annulus}
\alias{annulus}
\title{Plot an annulus on an existing plot.}
\usage{
annulus(x, y, inner, outer, deg = c(0, 360), n = 1000, col = "black",
  border = col, force0 = FALSE, ...)
}
\arguments{
\item{x}{Numeric, x-coordinate of center of annulus.}

\item{y}{Numeric, y-coordinate of center of annulus.}

\item{inner}{Numeric, inner radius.}

\item{outer}{Numeric, outer radius.}

\item{deg}{Two-element numeric list, degrees across which to draw the annulus (0 is "north" and increasing values go clockwise). Default is \code{c(0, 360)}.}

\item{n}{Integer, number of vertices used to approximate a circle.}

\item{col}{Character or integer, color used to fill annulus.  Note that if \code{col} and \code{border} are different colors and \code{deg} is \code{c(0, 360)} (the default), then you wil abel to see the line connecting the inner and outer borders.}

\item{border}{Character or integer, color used to draw annulus border.  Note that if \code{col} and \code{border} are different colors and \code{deg} is \code{c(0, 360)} (the default), then you wil abel to see the line connecting the inner and outer borders.}

\item{force0}{Logical, if \code{TRUE} then negative values of inner and outer are coerced to equal 0. If \code{FALSE} then throws an error.}

\item{...}{Arguments to send to \code{\link{graphics::polygon}}.}
}
\value{
Nothing (side-effect is to add an annulus to an existing plot).
}
\description{
This function adds an annulus to an existingg plot.
}
\examples{
plot(0, 0, xlim=c(-1, 1), ylim=c(-1, 1))
annulus(x=0.3, y=-0.2, inner=1, outer=0.6, col='cornflowerblue')
annulus(x=-0.3, y=0.2, inner=1, outer=0.6, deg=c(0, 135), col='firebrick1', border='black')
}
\seealso{
\code{\link{graphics::polygon}}
}
