% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art.r
\name{art}
\alias{art}
\title{Aligned rank transform of non-oaramaric data for further analysis using ANOVA}
\usage{
art(x, response = names(x)[1], factors = names(x)[2:ncol(x)],
  subject = NULL, fun = function(x) mean(x, na.rm = T), verbose = FALSE)
}
\arguments{
\item{x}{Data frame.}

\item{response}{Character. Names of column of \code{x} that has response variable (default is to use the first column).}

\item{factors}{Character list. Names of columns of \code{x} used to define factors and levels (default is to use all columns except for the first).}

\item{subject}{NULL or character. Name of column in \code{x} that has the subject variable. If NULL then this is ignored. If specified, residuals are calculated for each cell defined by factors, \emph{not} by subject and factors, but aligning is done using both factors and subject.}

\item{fun}{Function. Function used to calculate cell centering statistic (the default is to use: \code{mean} with \code{na.rm=TRUE}). The function can be any that handles a list of one or more elements.}

\item{verbose}{Logical. If TRUE then display progress.}
}
\value{
Data frame.
}
\description{
This function performs the aligned rank transforms on non-parametric data. This is useful for further analysis using parametric techniques like ANOVA.
}
\details{
The function successfully re-creates rankings given by \strong{ARTool} (Wobbrock et al. 2011) of data in Higgins et al. (1990) for data with 2 and 3 factors.
If \code{response} is ranks and the set of ranks in each cell is the same (e.g., each cell has ranks 1, 2, and 3, but not necessarily in that order), then all values will be equal across the different ART variables.  This occurs because the center of each cell (e.g., the mean) is the same as the grand mean, so the aligned values are simply the residuals. An ANOVA on this data yields no variance across cells, so the F tests are invalid.
}
\examples{
x <- data.frame(
   subject=c('a', 'b', 'c', 'a', 'b', 'c', 'a', 'b', 'c', 'a', 'b', 'c'),
   factor1=c('up', 'up', 'up', 'up', 'up', 'up', 'down', 'down', 'down', 'down',
      'down', 'down'),
   factor2=c('high', 'med', 'low', 'high', 'med', 'low', 'high', 'med', 'low', 'high',
      'med', 'low'),
   response=c(1, 17, 1, 1, 0, 4, 5, 6, 3, 7, 100, 70)
)
art(x=x, response='response', factors=c('factor1', 'factor2'))
}
\references{
Higgins, J.J., Blair, R.C., and Tashtoush, S.  1990.  The aligned rank transform procedure.  \emph{roceedings of the Conference on Applied Statistics in Agriculture.}  Manhattan, Kansas: Kansas State University, pp. 185-195.

Peterson, K.  2002.  Six modifications of the aligned rank transform test for interaction. \emph{Journal of Modern Applied Statistical Methods} 1:100-109.

Wobbrock, J.O., Findlater, L., Gergle, D., and Higgins, J.J.  2011. The aligned rank transform for nonparametric factorial analysis using only ANOVA procedures.  \emph{Proceedings of the ACM Conference on Human Factors in Computing Systems (CHI 2011).} Vancouver, British Columbia (May 7-12, 2011). New York: ACM Press, pp. 143-146.
}
