% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.r
\name{bracket}
\alias{bracket}
\title{Identify values bracketing another value}
\usage{
bracket(x, by, index = FALSE, warn = TRUE)
}
\arguments{
\item{x}{Numeric value or vector of values.}

\item{by}{A vector of numeric values. These should be sorted (from high to low or low to high... if not, an error will result).}

\item{index}{Logical. If \code{FALSE} (default), then numeric values in \code{by} will be returned. If \code{TRUE}, then the index or indices of the bracketing value(s) will be returned.}

\item{warn}{Logical. If \code{TRUE} (default), then warn if \code{x} is outside the range of \code{by}.}
}
\value{
If \code{x} is a single value, then the function will return a numeric vector of length 1 or 2, depending on how many values bracket \code{x}. If \code{x} is a vector, then the result will be a list with one element per item in \code{x} with each element having the same format as the case when \code{x} is a single value.
}
\description{
This function finds the set of values in a vector that bracket a third value, \code{x}. If \code{x} is exactly equal to one of the values in the vector, then a single value equal to \code{x} is returned. If \code{x} falls outside of the range of the vector, then the least/most extreme value of the vector is returned (depending on which side of the distribution of the vector \code{x} resides).  Optionally, users can have the function return the index of the values that bracket \code{x}.
}
\examples{
x <- 3.2
by <- 2 * (1:10)
bracket(3.2, by)
bracket(6.8, by)
bracket(18, by)

bracket(3.2, by, index=TRUE)

bracket(c(3.2, 9.8, 4), by)

bracket(2, c(0, 1, 1, 1, 3, 5), index=TRUE)
bracket(3, c(1, 2, 10))

\donttest{
by <- 1:10
bracket(-100, by)
bracket(100, by)
}

}
