% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversionFactors.r
\docType{data}
\name{conversionFactors}
\alias{conversionFactors}
\title{Data frame or conversion factors for length or areal units}
\format{
An object of class \code{'data.frame'}.
}
\usage{
conversionFactors
}
\description{
Data frame of conversion factors for length or areal units.
}
\examples{
data(conversionFactors)
conversionFactors
}
\seealso{
\code{\link{convertUnits}}, \code{\link{expandUnits}}
}
\keyword{datasets}
