% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cull.r
\name{cull}
\alias{cull}
\title{Force objects to have length or rows equal to the shortest}
\usage{
cull(...)
}
\arguments{
\item{...}{Vectors, matrices, or data frames.}
}
\description{
Force objects to have length or rows equal to the shortest
}
\examples{
a <- 1:10
b <- 1:20
c <- letters
cull(a, b, c)
x <- data.frame(x=1:10, y=letters[1:10])
y <- data.frame(x=1:26, y=letters)
cull(x, y)
}
