% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirCreate.r
\name{dirCreate}
\alias{dirCreate}
\title{Replacement for dir.create().}
\usage{
dirCreate(...)
}
\arguments{
\item{...}{Character string(s). These indicate the path of the directory to create. Multiple strings will be pasted together into one path.}
}
\value{
Nothing (creates a directory on the storage system).
}
\description{
This function is a somewhat friendier version of \code{\link[base]{dir.create}} in that it automatically sets \code{recursive=TRUE} and \code{showWarnings=FALSE} arguments.
}
\seealso{
\code{\link[base]{dir.create}}
}
