% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domLeap.r
\docType{data}
\name{domLeap}
\alias{domLeap}
\title{Day of month for leap years}
\format{An object of class \code{'data.frame'}.}
\source{

}
\usage{
data(doyLeap)
}
\description{
Data frame of day of month for each month in a leap year.
}
\examples{
data(domLeap)
domLeap
}
\references{

}
\keyword{datasets}
