% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domNonLeap.r
\docType{data}
\name{domNonLeap}
\alias{domNonLeap}
\title{Day of month for non-leap years}
\format{
An object of class \code{'data.frame'}.
}
\usage{
domNonLeap
}
\description{
Data frame of day of month for each month in a non-leap year.
}
\examples{
data(domNonLeap)
domNonLeap
}
\keyword{datasets}
