% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsos.r
\name{.ls.objects}
\alias{.ls.objects}
\title{Size of objects taking most memory}
\usage{
.ls.objects(n, pattern, orderBy, decreasing = FALSE, pos = 1)
}
\arguments{
\item{n}{Positive integer, number of objects to display.}

\item{pattern}{Character, return size of objects with this pattern in their name.}

\item{orderBy}{Either \code{NULL} (default) or any of c\code{'Type'}, \code{'Size'}, \code{'PrettySize'}, \code{'Rows'}, or \code{'Columns'}. Controls field by which objects are sorted.}

\item{decreasing}{Logical, if \code{FALSE} (default), objects are displayed from larges to smallest.}

\item{pos}{Integer, indicates environment in which to search for objects. See help for \code{\link{ls}}.}
}
\value{
Nothing (displays a table).
}
\description{
Displays the largest objects in memory.
}
\details{
Adapted from StackOverflow user Dirk Eddelbuettel (original function at https://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session).
}
\examples{
lsos()
lsos(3)
}
