% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doyLeap.r
\docType{data}
\name{doyLeap}
\alias{doyLeap}
\title{Day of year for leap years}
\format{
An object of class \code{'data.frame'}.
}
\usage{
doyLeap
}
\description{
Data frame of day of year for each month in a leap year.
}
\examples{
data(doyLeap)
doyLeap
}
\keyword{datasets}
