% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doyNonLeap.r
\docType{data}
\name{doyNonLeap}
\alias{doyNonLeap}
\title{Day of year for non-leap years}
\format{An object of class \code{'data.frame'}.}
\source{

}
\usage{
data(doyNonLeap)
}
\description{
Data frame of day of year for each month in a non-leap year
}
\examples{
data(doyNonLeap)
doyNonLeap
}
\references{

}
\keyword{datasets}
