% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipseNames.r
\name{ellipseNames}
\alias{ellipseNames}
\title{Get names of objects input as arguments in ellipse (\code{...}) form}
\usage{
ellipseNames(...)
}
\arguments{
\item{...}{Objects.}
}
\value{
Character list.
}
\description{
This function returns the names of objects input into a function as ellipses. It is only useful if called inside a function.
}
