% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps.r
\name{eps}
\alias{eps}
\title{The smallest machine-readable number}
\usage{
eps()
}
\value{
Numeric value.
}
\description{
This function returns the smallest machine-readable number (equal to \code{.Machine$double.eps}).
}
\examples{
eps()
}
