% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertCol.r
\name{insertCol}
\alias{insertCol}
\title{Insert a column into a data frame or matrix}
\usage{
insertCol(x, into, at, before = TRUE)
}
\arguments{
\item{x}{Data frame, matrix, or vector with same number of rows or elements as \code{into}.}

\item{into}{Data frame or matrix into which \code{x} is to be inserted.}

\item{at}{Character or integer, name of column or column number at which to do insertion.}

\item{before}{Logical, if \code{TRUE} (default) then the insertion will occur in front of the column named in \code{at}, if \code{FALSE} then after.}
}
\value{
A data frame.
}
\description{
This function inserts one or more columns before or after another column in a data frame or matrix. It is similar to \code{\link[base]{cbind}} except that the inserted column(s) can be placed anywhere.
}
\examples{
x <- data.frame(y1=11:15, y2=rev(letters)[1:5])
into <- data.frame(x1=1:5, x2='valid', x3=letters[1:5], x4=LETTERS[1:5], x5='stuff')

insertCol(x, into=into, at='x3')
insertCol(x, into=into, at='x3', before=FALSE)
}
\seealso{
\code{\link[base]{merge}}, \code{\link[base]{cbind}}
}
