% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertRow.r
\name{insertRow}
\alias{insertRow}
\title{Insert a row into a data frame or matrix}
\usage{
insertRow(x, into, at = NULL, before = TRUE)
}
\arguments{
\item{x}{Data frame, matrix, or vector with same number of columns or elements as \code{into}.}

\item{into}{Data frame or matrix into which \code{x} is to be inserted.}

\item{at}{Character, integer, or \code{NULL}. Name of row or row number at which to do insertion. If \code{NULL} (default), the result is exactly the same as \code{rbind(into, x}.}

\item{before}{Logical, if \code{TRUE} (default) then the insertion will occur immediately above the row named in \code{at}, if \code{FALSE} then after. Ignored if \code{at} is \code{NULL}.}
}
\value{
A data frame.
}
\description{
This function inserts one or more rows before or after another row in a data frame or matrix. It is similar to \code{\link{rbind}} except that the inserted rows(s) can be placed anywhere.
}
\examples{
x <- data.frame(x1=1:3, x2=LETTERS[1:3])
into <- data.frame(x1=11:15, x2='valid')
row.names(into) <- letters[1:5]

insertRow(x, into=into, at='b')
insertRow(x, into=into, at='b', before=FALSE)
insertRow(x, into)
}
\seealso{
\code{\link[base]{merge}}, \code{\link{rbind}}, \code{\link{insertCol}}
}
