% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isLeapYear.r
\name{isLeapYear}
\alias{isLeapYear}
\title{Is a year a leap year?}
\usage{
isLeapYear(x)
}
\arguments{
\item{x}{Integer or vector of integers representing years.}
}
\value{
Vector of logical values.
}
\description{
Returns \code{TRUE} if the year is a leap year. You can use "negative" years for BCE.
}
\examples{
isLeapYear(1990:2004) # note 2000 *was* not a leap year
isLeapYear(1896:1904) # 1900 was *not* a leap year
}
