% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFiles.r
\name{listFiles}
\alias{listFiles}
\title{Replacement for \code{list.files()}}
\usage{
listFiles(x, ...)
}
\arguments{
\item{x}{Path name of folder containing files to list.}

\item{...}{Arguments to pass to \code{list.files} (other than \code{full.names}).}
}
\value{
Character.
}
\description{
This function is a slightly friendlier version of \code{\link[base]{list.files}} in that it automatically includes the \code{full.names=TRUE} argument.
}
\examples{
# list files in location where R is installed
listFiles(R.home())
listFiles(R.home(), pattern='README')
}
\seealso{
\code{\link[base]{list.files}}
}
