% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsos.r
\name{lsos}
\alias{lsos}
\title{Size of objects taking most memory}
\usage{
lsos(n = 10, orderBy = "size", decreasing = TRUE, pos = 1, ...)
}
\arguments{
\item{n}{Positive integer, maximum number of objects to display.}

\item{orderBy}{Either \code{'size'} (default) or \code{'name'}.}

\item{decreasing}{Logical, if \code{TRUE} (default), objects are displayed from largest to smallest.}

\item{pos}{Environment from which to obtain size of objects. Default is 1. See \code{\link{ls}}.}

\item{...}{Other arguments to pass to \code{\link{ls}}.}
}
\value{
Nothing (displays a table).
}
\description{
Displays the largest objects in memory.
}
\details{
Adapted from StackOverflow user JD Long (original function at https://stackoverflow.com/questions/1395270/determining-memory-usage-of-objects).
}
\examples{
lsos()
lsos(3)
}
