% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memUse.r
\name{memUse}
\alias{memUse}
\title{Size of objects taking most memory use}
\usage{
memUse(
  n = 10,
  orderBy = "size",
  decreasing = TRUE,
  pos = 1,
  display = TRUE,
  ...
)
}
\arguments{
\item{n}{Positive integer, maximum number of objects to display.}

\item{orderBy}{Either \code{'size'} (default) or \code{'name'}.}

\item{decreasing}{Logical, if \code{TRUE} (default), objects are displayed from largest to smallest.}

\item{pos}{Environment from which to obtain size of objects. Default is 1. See \code{\link{ls}}.#"}

\item{display}{If \code{TRUE} (default), print a table with memUse used.}

\item{...}{Other arguments to pass to \code{\link{ls}}.}
}
\value{
Data frame (invisible).
}
\description{
Displays the largest objects in memUse.
}
\examples{
memUse()
memUse(3)
}
