% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmode.r
\name{mmode}
\alias{mmode}
\title{Calculate the mode of numeric, character, or factor data}
\usage{
mmode(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric, character, or factor vector.}

\item{na.rm}{Logical. If \code{TRUE} then remove \code{NA}s first. Otherwise fail.}
}
\value{
Numeric, character, or factor value.
}
\description{
Calculate the mode of numeric, character, or factor data
}
\details{
Adapted from Ken Williams' post on StackExchange [https://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode].
}
\examples{
mmode(round(10 * rnorm(1000, 2)))
mmode(c('a', 'b', 'b', 'b', 'c', 'd', 'd'))
}
