% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naRows.r
\name{naRows}
\alias{naRows}
\title{Index any rows in a data frame or matrix that contain at least one \code{NA}}
\usage{
naRows(x)
}
\arguments{
\item{x}{Data frame or matrix.}
}
\value{
Integer vector.
}
\description{
This function returns the row number of any row in a data frame or matrix that has at least one \code{NA}.
}
\examples{
x <- data.frame(a=1:5, b=c(1, 2, NA, 4, 5), c=c('a', 'b', 'c', 'd', NA))
naRows(x)
}
