% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairDist.r
\name{pairDist}
\alias{pairDist}
\title{Calculate pairwise distances between two matrices or data frames}
\usage{
pairDist(set1, set2, na.rm = FALSE)
}
\arguments{
\item{set1}{Data frame or matrix one or more columns wide.}

\item{set2}{Data frame or matrix one or more columns wide.}

\item{na.rm}{Logical, if \code{TRUE} then any rows in \code{set1} or \code{set2} with at least one \code{NA} are removed first.}
}
\value{
Matrix with \code{nrow(set1)} rows and \code{nrow(set2)} columns.
}
\description{
This function takes two data frames or matrices and returns a matrix of Euclidean distances between the two.
}
\examples{
set1 <- data.frame(x1=1:20, x2=round(100 * rnorm(20)))
set2 <- data.frame(x1=sample(1:30, 30), x2=sort(round(100 * rnorm(30))))
pairDist(set1, set2)
}
