% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probitAdj.r
\name{probitAdj}
\alias{probitAdj}
\title{The inverse of the [logitAdj()] function which is robust to cases that equal 0 or 1}
\usage{
probitAdj(x, epsilon = 0.01, base = 10)
}
\arguments{
\item{x}{Numeric list.}

\item{epsilon}{Value to add/subtract from x to ensure log of 0 or 1 is not taken (usually a small number).}

\item{base}{Base of logarithm. Use \code{base=exp(1)} for base \code{e}.}
}
\value{
Numeric.
}
\description{
This function is the inverse of the \code{logitAdj()}. That function calcualtes the logit of values but is robust to cases where the operadn is 0 or 1. The adjusted probit is equal to \code{(base^x + epsilon * base^x - epsilon) / (base^x + 1)}.
}
\examples{
set.seed(123)
x <- seq(0, 1, by=0.01)
logitAdj(x)
logitAdj(x, 0.001)
logitAdj(x, base=2)
}
\seealso{
[logitAdj()]
}
