% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadArea.r
\name{quadArea}
\alias{quadArea}
\title{Area of a quadrilateral}
\usage{
quadArea(x, y)
}
\arguments{
\item{x}{Numeric list. \code{x} coordinates of quadrilateral.}

\item{y}{Numeric list. \code{y} coordinates of quadrilateral.}
}
\value{
Numeric (area of a quadrilateral in same units as \code{x} and \code{y}.
}
\description{
Calculates the area of a quadrilaterial by dividing the quadrilateral into two triangles.
}
\examples{
x <- c(0, 6, 4, 1)
y <- c(0, 1, 7, 4)
quadArea(x, y)
plot(1, type='n', xlim=c(0, 7), ylim=c(0, 7), xlab='x', ylab='y')
polygon(x, y)
}
