% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundTo.r
\name{roundTo}
\alias{roundTo}
\title{Round to nearest target value}
\usage{
roundTo(x, target, roundFx = round)
}
\arguments{
\item{x}{Numeric.}

\item{target}{Numeric.}

\item{roundFx}{Any of \code{\link{round}}, \code{\link{floor}}, or \code{\link{ceiling}}.}
}
\value{
Numeric.
}
\description{
This function rounds a value to a nearest "target" value (e.g., you could round 0.72 to the nearest 0.25, or 0.75).
}
\examples{
roundTo(0.73, 0.05)
roundTo(0.73, 0.1)
roundTo(0.73, 0.25)
roundTo(0.73, 0.25, floor)
roundTo(0.73, 1)
roundTo(0.73, 10)
roundTo(0.73, 10, ceiling)
}
