% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same.r
\name{same}
\alias{same}
\title{Tests if all elements of a vector are the same.}
\usage{
same(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric or character vector.}

\item{na.rm}{Logical, if \code{TRUE} then remove \code{NA}s first. If \code{FALSE} and \code{NA}s are included with other values, then the function returns \code{FALSE}.}
}
\description{
Tests if all elements of a vector are the same.
}
\examples{
same(1:10)
same(rep(3, 10))
same(letters[1:3])
same(c('r', 'r', 'r'))
same(c('r', 'r', 'r', NA))
same(c('r', 'r', 'r', NA), na.rm=TRUE)
}
