% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenRes.r
\name{screenRes}
\alias{screenRes}
\title{Find screen resolution}
\usage{
screenRes()
}
\value{
A list with two elements: `width` and `height`, which are the width and height of the screen in pixels.
}
\description{
This function finds the screen resolution of the current device. It should work on Windows, macOS, and Linux systems.
}
\examples{
screenRes()

}
