% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.r
\name{se}
\alias{se}
\title{Standard error}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric list.}

\item{na.rm}{Logical. If TRUE then remove \code{NA}s before calculation.}
}
\value{
Numeric.
}
\description{
Calculate the standard error of the mean.
}
\examples{
se(1:100)
}
\seealso{
[stats::sd()]
}
