% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlistRecursive.r
\name{unlistRecursive}
\alias{unlistRecursive}
\title{For any object in a list that is also a list, unlist it}
\usage{
unlistRecursive(x)
}
\arguments{
\item{x}{A \code{list}.}
}
\value{
A \code{list}.
}
\description{
This function takes as an argument a \code{list}. If any of its elements are also lists, it unlists them. The output is the same as the input, except that there will be one new element per element in each sublist, and the sublists will be removed.
}
\examples{

x <- list(
   a = 1:3,
   b = list(
      b1 = c("The", "quick", "brown", "function"),
	     b2 = 4:1,
      b3 = list(
         b3_1 = 5:7
      )
   ),
   c = "end"
)

unlistRecursive(x)

}
\seealso{
\code{\link{unlist}}
}
