% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsink.r
\name{unsink}
\alias{unsink}
\title{Close all open calls of sink()}
\usage{
unsink(x = 10, warn = TRUE)
}
\arguments{
\item{x}{Numeric: Number of sink files to try to close.}

\item{warn}{Logical: If \code{TRUE}, display warnings if attempts are made to close sink files that are not open.}
}
\value{
Logical (invisibly).
}
\description{
If you use the [sink()] function frequently, you can inadvertently open files while debugging. This function closes all open \code{sink} files.
}
\examples{

# create 3 sink files
for (i in 1:3) {
   tempFile <- tempfile(fileext = '.txt')
   sink(tempFile, split = TRUE)
   say('We will print this message on the screen and to a file.')
}
unsink(3) # close all 3

}
