% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirCreate.r
\name{dirCreate}
\alias{dirCreate}
\title{Replacement for dir.create()}
\usage{
dirCreate(...)
}
\arguments{
\item{...}{Character string(s). The path and name of the directory to create. Multiple strings will be pasted together into one path, although slashes will not be pasted between them.}
}
\value{
Nothing (creates a directory on the storage system).
}
\description{
This function is a somewhat friendlier version of \code{\link{dir.create}} in that it automatically sets \code{recursive=TRUE} and \code{showWarnings=FALSE} arguments.
}
\seealso{
\code{\link{dir.create}}
}
