% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doyNonLeap.r
\docType{data}
\name{doyNonLeap}
\alias{doyNonLeap}
\title{Days of year for non-leap years}
\format{
An object of class \code{'data.frame'}.
}
\usage{
doyNonLeap
}
\description{
Data frame of days of year for each month in a non-leap year
}
\examples{
data(doyNonLeap)
doyNonLeap
}
\keyword{datasets}
