% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmode.r
\name{mmode}
\alias{mmode}
\title{Calculate the mode of numeric, character, or factor data.}
\usage{
mmode(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric, character, or factor vector.}

\item{na.rm}{Logical. If \code{TRUE} then remove \code{NA}s first. Otherwise fail.}
}
\value{
Numeric, character, or factor value.
}
\description{
Calculate the mode of numeric, character, or factor data.
}
