% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmode.r
\name{mmode}
\alias{mmode}
\title{Modal value(s)}
\usage{
mmode(x)
}
\arguments{
\item{x}{Numeric or character vector.}
}
\value{
Numeric or character vector.
}
\description{
Modal value. If there is more than one unique mode, all modal values are returned.
}
\examples{
x <- c(1, 2, 3, 3, 4, 5, 3, 1, 2)
mmode(x)

x <- c(1, 2, 3)
mmode(x)

}
