% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naCompare.r
\name{naCompare}
\alias{naCompare}
\title{Compare numeric values using <, <=, >, >=, !=, and == robustly to NAs}
\usage{
naCompare(op, x, y)
}
\arguments{
\item{op}{Character, the operation to perform: \code{'<'}, \code{'<='}, \code{'>'}, \code{'>='}, \code{'!='}, or \code{'=='}. Note this must be a character (i.e., put it in quotes).}
}
\value{
Vector of logical values.
}
\description{
This function performs simple (vectorized) comparisons using \code{<}, \code{<=}, \code{>}, \code{>=}, \code{!=}, or \code{==}, between values and \emph{always} returns \code{TRUE} or \code{FALSE}. \code{TRUE} only occurs if the condition can be evlaauted and it is \code{TRUE}. \code{FALSE} is returned if the condition cannot be evaluated \emph{or} it is \code{FALSE}. Normally, if the condition cannot be evaluated (e.g., because at least one of the values is \code{NA} or \code{NaN}), then \code{NA} is returned.
}
\examples{
naCompare('<', c(1, 2, NA), c(10, 1, 0))
naCompare('<', c(1, 2, NA), 10)
naCompare('<', c(1, 2, NA), NA)
# compare to:
NA < 5
NA < NA
}
