% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notIn.r
\name{notIn}
\alias{notIn}
\alias{`\%notin\%`}
\alias{\%notin\%}
\title{Opposite of '\%in\%'}
\usage{
notIn(x, table)

x \%notin\% table
}
\arguments{
\item{x, table}{Vectors.}
}
\value{
A logical vector.
}
\description{
Indicate if elements of a vector are not in another vector.
}
\examples{

x <- c('a', 'v', 'o', 'C', 'a', 'd', 'O')
y <- letters

y \%notin\% x
x \%notin\% y

}
