% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnibus.r
\docType{package}
\name{omnibus}
\alias{omnibus}
\alias{omnibus-package}
\title{omnibus: Fantabulous helper functions}
\description{
This package contains a set of helper functions.
}
\details{
Create an issue on \href{https://github.com/adamlilith/omnibus/issues}{GitHub}.
}
\author{
Adam B. Smith
}
